WallMessages/wallapi/requirements.txt
abstract-rendering==0.5.1
alabaster==0.7.3
appnope==0.1.0
appscript==1.0.1
argcomplete==0.8.9
asn1crypto==0.22.0
astropy==1.0.4
Babel==1.3
bcolz==0.9.0
beautifulsoup4==4.3.2
binstar==0.11.0
bitarray==0.8.1
blaze==0.8.0
blz==0.6.2
bokeh==0.9.0
boto==2.38.0
Bottleneck==1.0.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
clyent==0.3.4
colorama==0.3.3
conda==4.3.25
conda-build==1.14.1
configobj==5.0.6
cryptography==1.8.1
Cython==0.22.1
cytoolz==0.7.3
datashape==0.4.5
decorator==4.0.4
Django==1.11.3
django-appconf==1.0.2
django-filter==1.1.0
Django-Select2==5.11.1
django-suit==0.2.25
djangorestframework==3.6.4
docutils==0.12
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.10.1
glueviz==0.5.3
greenlet==0.4.7
h5py==2.5.0
idna==2.5
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
jdcal==1.3
jedi==0.8.1
Jinja2==2.7.3
jsonschema==2.4.0
jupyter-client==4.1.1
jupyter-core==4.0.6
llvmlite==0.5.0
lxml==3.4.4
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.5.1
mock==1.0.1
multipledispatch==0.4.7
mysqlclient==1.3.12
networkx==1.10
nltk==3.0.3
nose==1.3.7
numba==0.19.1
numexpr==2.4.3
numpy==1.9.3
odo==0.3.2
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.16.2
path.py==0.0.0
patsy==0.3.0
pep8==1.6.2
pexpect==3.3
pickleshare==0.5
Pillow==4.2.1
ply==3.6
psutil==2.2.1
ptyprocess==0.4
py==1.4.27
pyasn1==0.1.7
pycosat==0.6.2
pycparser==2.18
pycrypto==2.6.1
pycurl==7.19.5.1
pyfcm==1.4.2
pyflakes==0.9.2
Pygments==2.0.2
pyOpenSSL==17.0.0
pyparsing==2.2.0
pytest==2.7.1
python-dateutil==2.4.2
pytz==2017.2
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.1.0
redis==2.10.3
requests==2.18.1
requests-toolbelt==0.8.0
rope-py3k==0.9.4.post1
runipy==0.1.3
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.16.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spyder==2.3.5.2
SQLAlchemy==1.0.5
statsmodels==0.6.1
sympy==0.7.6
tables==3.2.0
terminado==0.5
toolz==0.7.2
tornado==4.2
traitlets==4.0.0
ujson==1.33
urllib3==1.21.1
virtualenv==15.1.0
Werkzeug==0.10.4
whitenoise==2.0
xlrd==0.9.3
XlsxWriter==0.7.3
xlwings==0.3.5
xlwt==1.0.0
